requirements.txt
Django==1.8.5
django-modeltranslation==0.10.1
djangorestframework==3.2
django-reversion==1.9.3
django-cors-headers==1.1.0
django-filter==0.11.0
pytest==2.8.2
pytest_django==2.9.1
pillow==3.0.0
easy-thumbnails==2.2
pytest-cov==2.2.0
flake8==2.4.1
factory-boy==2.6.0
drf-nested-routers==0.10.0
xlsxwriter==0.7.7
git+https://github.com/City-of-Helsinki/django-helusers.git@405028e4c45a33e0aaf893fee5#egg=django-helusers
djangorestframework-jwt
pyjwt
django-enumfields==0.7.4
